Refactored namespace and added TokenBinder
The namespace for Docker related classes has been changed from 'UnMango.Tdl.Cli.Internal' to 'UnMango.Tdl.Cli.Docker'. This change improves the organization of the code by grouping Docker-related functionality under a dedicated namespace. Additionally, a new class named 'TokenBinder' has been introduced in the 'UnMango.Tdl.Cli.Internal' namespace. This class is responsible for binding CancellationToken instances.
