+++ src/Cli/Docker/Container.cs
@@ -1,7 +1,7 @@
 using Docker.DotNet.Models;
 using Serilog;
 
-namespace UnMango.Tdl.Cli.Internal;
+namespace UnMango.Tdl.Cli.Docker;
 
 public interface IContainer : IAsyncDisposable
 {
