+++ src/Cli/Docker/IDockerProgress.cs
@@ -1,7 +1,6 @@
-using System.CommandLine;
 using Docker.DotNet.Models;
 
-namespace UnMango.Tdl.Cli.Internal;
+namespace UnMango.Tdl.Cli.Docker;
 
 internal interface IDockerProgress : IProgress<string>, IProgress<JSONMessage>;
