+++ src/Cli/Docker/SerilogProgress.cs
@@ -1,7 +1,7 @@
 using Docker.DotNet.Models;
 using Serilog;
 
-namespace UnMango.Tdl.Cli.Internal;
+namespace UnMango.Tdl.Cli.Docker;
 
 internal sealed class SerilogProgress : IDockerProgress
 {
